Feat: Se generan componenetes para la pagina main.
