+++ SMPFront/src/pages/MainPage.jsx
@@ -4,6 +4,7 @@
 import { useState } from "react"
 import { Search, Home, BookOpen, TrendingUp, Award, Bell, MessageCircle, User, ChevronRight } from 'lucide-react'
 import "../styles/MainPage.css"
+import { ModuleCard } from "../components/Modules/Modulecard"
 
 export default function MainPage() {
   const [activeSection, setActiveSection] = useState("inicio")
@@ -98,153 +99,11 @@ export default function MainPage() {
     },
   ]
 
-  const sidebarItems = [
-    { id: "inicio", label: "Inicio", icon: Home },
-    { id: "mis-rutas", label: "Mis Rutas", icon: BookOpen },
-    { id: "mi-progreso", label: "Mi Progreso", icon: TrendingUp },
-    { id: "mis-certificados", label: "Mis Certificados", icon: Award },
-    { id: "notificaciones", label: "Notificaciones", icon: Bell },
-  ]
-
   return (
-    <div className="main-page-container">
-      {/* Header */}
-      <header className="main-header">
-        <div className="header-content">
-          <div className="header-left">
-            <div className="logo-section">
-              <div className="logo-icon">
-                {logoUrl ? (
-                  <img 
-                    src={logoUrl || "/placeholder.svg"} 
-                    alt="Semillas de Palabras Logo" 
-                    className="logo-image"
-                    onError={(e) => {
-                      // Si la imagen no carga, muestra el texto SP
-                      e.target.style.display = 'none'
-                      e.target.nextSibling.style.display = 'block'
-                    }}
-                  />
-                ) : null}
-                <span className="logo-fallback">SP</span>
-              </div>
-              <span className="logo-text">Semillas de Palabras</span>
-            </div>
-            <div className="search-section">
-              <div className="search-icon">
-                <Search size={16} />
-              </div>
-              <input placeholder="¿Qué quieres aprender?" className="search-input" />
-            </div>
-          </div>
-          <div className="header-right">
-            <button className="header-button">
-              <Bell size={16} />
-              <span>0</span>
-            </button>
-            <button className="header-button">Planes</button>
-            <div className="user-section">
-              <div className="user-avatar">
-                <User size={16} />
-              </div>
-              <span>500 pts</span>
-            </div>
-          </div>
-        </div>
-      </header>
-
-      <div className="main-layout">
-        {/* Sidebar */}
-        <aside className="sidebar">
-          <nav className="sidebar-nav">
-            <ul className="sidebar-list">
-              {sidebarItems.map((item) => {
-                const Icon = item.icon
-                return (
-                  <li key={item.id}>
-                    <button
-                      onClick={() => setActiveSection(item.id)}
-                      className={`sidebar-item ${activeSection === item.id ? "active" : ""}`}
-                    >
-                      <Icon size={20} />
-                      <span>{item.label}</span>
-                    </button>
-                  </li>
-                )
-              })}
-            </ul>
-          </nav>
-        </aside>
-
-        {/* Main Content */}
-        <main className="main-content">
-          {/* Welcome Section */}
-          <section className="welcome-section">
-            <h1 className="welcome-title">Descubre lo que puedes aprender</h1>
-            <p className="welcome-subtitle">Cada palabra es una semilla que puede crecer en conocimiento</p>
-          </section>
-
-          {/* Recommended Courses */}
-          <section className="courses-section">
-            <div className="section-header">
-              <h2 className="section-title">Cursos recomendados para ti</h2>
-              <button className="view-all-button">
-                Ver todos <ChevronRight size={16} />
-              </button>
-            </div>
-            <div className="courses-grid">
-              {recommendedCourses.map((course) => (
-                <div key={course.id} className="course-card">
-                  <div className="course-image-container">
-                    <img src={course.image || "/placeholder.svg"} alt={course.title} className="course-image" />
-                    {course.isNew && <span className="new-badge">NUEVO</span>}
-                  </div>
-                  <div className="course-content">
-                    <div className="course-category">
-                      <div className="category-icon">
-                        <BookOpen size={12} />
-                      </div>
-                      <span>{course.category}</span>
-                    </div>
-                    <h3 className="course-title">{course.title}</h3>
-                    <p className="course-instructor">Por {course.instructor}</p>
-                  </div>
-                </div>
-              ))}
-            </div>
-          </section>
-
-          {/* Latest Courses */}
-          <section className="courses-section">
-            <div className="section-header">
-              <h2 className="section-title">Últimos cursos lanzados</h2>
-              <button className="view-all-button">
-                Ver todos <ChevronRight size={16} />
-              </button>
-            </div>
-            <div className="courses-grid">
-              {latestCourses.map((course) => (
-                <div key={course.id} className="course-card">
-                  <div className="course-image-container">
-                    <img src={course.image || "/placeholder.svg"} alt={course.title} className="course-image" />
-                    {course.isNew && <span className="new-badge">NUEVO</span>}
-                  </div>
-                  <div className="course-content">
-                    <div className="course-category">
-                      <div className="category-icon category-icon-alt">
-                        <BookOpen size={12} />
-                      </div>
-                      <span>{course.category}</span>
-                    </div>
-                    <h3 className="course-title">{course.title}</h3>
-                    <p className="course-instructor">Por {course.instructor}</p>
-                  </div>
-                </div>
-              ))}
-            </div>
-          </section>
-        </main>
-      </div>
-    </div>
+   <>
+   
+   <ModuleCard Modules={recommendedCourses} ModulosRecientes={latestCourses}/>
+   
+   </>
   )
 }
