+++ SMPFront/src/pages/MainPages/PageContainer.jsx
@@ -1,0 +1,17 @@
+import React from "react"
+import { Outlet } from 'react-router-dom' 
+import { NavBars } from "../../components/MainPage/NavBars"
+
+
+export const Mainpages = ()=>{
+
+
+    return(<>
+    
+        <NavBars/>
+        <Outlet/>
+    
+    </>
+    )
+
+}
